requirements.txt
black==22.12.0
Django==4.1.5
flake8==6.0.0
isort==5.11.4
pytest-mypy==0.10.3
